refactor(phase0): 不要ファイルの削除と補助スクリプトの整理
- 非推奨スクリプトを削除: train_controlnet.py, sdxl_train_control_net_lllite_old.py,
  networks/merge_lora_old.py, train_textual_inversion_XTI.py
- 不要ディレクトリを削除: bitsandbytes_windows/, wuerstchen/
- XTI_hijack.py を gen_img_diffusers.py にインライン化して削除
- pytest 用 5–6 行スタブテストを削除（test_train.py 等 10 件）
- 開発補助スクリプトを tools/dev/ へ移動: manual_test_anima_*.py, visualize_masks.py

インペインティングテストランナー (run_*_inpainting_test.*) と密結合する
_verify_inpainting_checkpoint.py / generate_inpainting_test_data.py /
download_training_data.py は tests/ に維持。
