--- tests/test_fine_tune.py
@@ -1,6 +1,0 @@
-import fine_tune
-
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
