--- tests/test_flux_train.py
@@ -1,6 +1,0 @@
-import flux_train
-
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
