--- tests/test_sd3_train.py
@@ -1,6 +1,0 @@
-import sd3_train
-
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
