--- tests/test_sd3_train_network.py
@@ -1,5 +1,0 @@
-import sd3_train_network
-
-def test_syntax():
-    # Very simply testing that the flux_train_network imports without syntax errors
-    assert True
