--- tests/test_sdxl_train_network.py
@@ -1,6 +1,0 @@
-import sdxl_train_network
-
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
