--- tests/test_train_network.py
@@ -1,5 +1,0 @@
-import train_network
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
