--- tests/test_train_textual_inversion.py
@@ -1,5 +1,0 @@
-import train_textual_inversion
-
-def test_syntax():
-    # Very simply testing that the train_network imports without syntax errors
-    assert True
