+++ tools/dev/manual_test_anima_real_training.py
